Refactor & Cleanup
