+++ src/DependencyInjection/BitzArt.TransientServiceProvider/BitzArt.TransientServiceProvider.csproj
@@ -22,7 +22,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.2" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.8" />
+		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.8" />
 	</ItemGroup>
 
 	<ItemGroup>
